Initial commit on the ribbon-style of programming.
Added redux for better stage management as well as persistent state.

Blocks can be added, removed and dragged to switch places. They are ordered by index and their position is saved in the persistent Redux state.

(Note: Some of the functionality of the old code canvas (D-Pad etc.) is still in this commit in case we will need it later on).